made @state, @emits block local in DSL
Deleted state related specs as the state is not observable from the API.
As such we don't need specs for them. What matters is what user actions fire.

This also fixes a bug, where the @ was missing on @state. The specs
defined attr_reader for the Example class, which made the specs work.
The culprit here was that the DSL implementation leaked out to the user
blocks.